requirements.txt
web3~=5.13.1
PyYAML~=5.3.1
matplotlib~=3.3.4
Flask~=2.0.1
requests_toolbelt
eciespy
daemoniker~=0.2.3
requests~=2.25.1
torch~=1.8.1+cu111
ipfshttpclient~=0.7.0a1
numpy~=1.20.2
networkx~=2.5.1
torchvision~=0.9.1+cu111
ipython>=6.0
zcommons
traitlets~=5.0.5
hjson~=3.0.2
setuptools~=56.0.0